refactor: migrate UI from monolothic to modular
