+++ assets/images/src/matplotlib/matplotlib_large.png
