+++ assets/images/src/matplotlib/move_large.png
