+++ assets/images/src/matplotlib/qt4_editor_options_large.png
