+++ assets/images/src/matplotlib/subplots_large.png
