+++ assets/images/src/matplotlib/zoom_to_rect_large.png
